requirements.txt
PyMuPDF==1.24.1
Pillow==10.4.0
